requirements.txt
fastapi
uvicorn[standard]
httpx
cachetools
jinja2
python-dotenv
